Player cog + checks
