+++ discord_bots/commands.py
@@ -54,6 +54,9 @@ from discord_bots.utils import (
     update_next_map_to_map_after_next,
     upload_stats_screenshot_imgkit_channel,
     win_probability,
+    is_in_game,
+    get_player_game,
+    finished_game_str,
 )
 
 from .bot import bot
@@ -91,6 +94,7 @@ from .models import (
 from .names import generate_be_name, generate_ds_name
 from .queues import AddPlayerQueueMessage, add_player_queue, waitlist_messages
 from .twitch import twitch
+
 
 _log = logging.getLogger(__name__)
 
@@ -487,7 +491,7 @@ async def create_game(
             )
         await asyncio.gather(*send_message_coroutines)
 
-        in_progress_game_cog = bot.get_cog("InProgressGameCog")
+        in_progress_game_cog = bot.get_cog("InProgressGameCommands")
         if (
             in_progress_game_cog is not None
             and isinstance(in_progress_game_cog, InProgressGameCommands)
@@ -498,7 +502,7 @@ async def create_game(
             )
             game.message_id = message.id
         else:
-            _log.warning("Could not get InProgressGameCog")
+            _log.warning("Could not get InProgressGameCommands")
 
         session.query(QueuePlayer).filter(QueuePlayer.player_id.in_(player_ids)).delete()  # type: ignore
         session.commit()
@@ -764,94 +768,6 @@ def mock_finished_game_teams_str(
         return output
 
 
-def finished_game_str(finished_game: FinishedGame, debug: bool = False) -> str:
-    """
-    Helper method to pretty print a finished game
-    """
-    output = ""
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        short_game_id = short_uuid(finished_game.game_id)
-        team0_fg_players: list[FinishedGamePlayer] = (
-            session.query(FinishedGamePlayer)
-            .filter(
-                FinishedGamePlayer.finished_game_id == finished_game.id,
-                FinishedGamePlayer.team == 0,
-            )
-            .all()
-        )
-        team1_fg_players: list[FinishedGamePlayer] = (
-            session.query(FinishedGamePlayer)
-            .filter(
-                FinishedGamePlayer.finished_game_id == finished_game.id,
-                FinishedGamePlayer.team == 1,
-            )
-            .all()
-        )
-
-        if config.SHOW_TRUESKILL:
-            output += f"**{finished_game.queue_name}** - **{finished_game.map_short_name}** ({short_game_id}) (mu: {round(finished_game.average_trueskill, 2)})"
-        else:
-            output += f"**{finished_game.queue_name}** - **{finished_game.map_short_name}** ({short_game_id})"
-
-        team0_player_ids = set(map(lambda x: x.player_id, team0_fg_players))
-        team1_player_ids = set(map(lambda x: x.player_id, team1_fg_players))
-        team0_fgp_by_id = {fgp.player_id: fgp for fgp in team0_fg_players}
-        team1_fgp_by_id = {fgp.player_id: fgp for fgp in team1_fg_players}
-        team0_players: list[Player] = session.query(Player).filter(Player.id.in_(team0_player_ids))  # type: ignore
-        team1_players: list[Player] = session.query(Player).filter(Player.id.in_(team1_player_ids))  # type: ignore
-        if debug:
-            team0_names = ", ".join(
-                sorted(
-                    [
-                        f"{escape_markdown(player.name)} ({round(team0_fgp_by_id[player.id].rated_trueskill_mu_before, 1)})"
-                        for player in team0_players
-                    ]
-                )
-            )
-            team1_names = ", ".join(
-                sorted(
-                    [
-                        f"{escape_markdown(player.name)} ({round(team1_fgp_by_id[player.id].rated_trueskill_mu_before, 1)})"
-                        for player in team1_players
-                    ]
-                )
-            )
-        else:
-            team0_names = ", ".join(
-                sorted([escape_markdown(player.name) for player in team0_players])
-            )
-            team1_names = ", ".join(
-                sorted([escape_markdown(player.name) for player in team1_players])
-            )
-        team0_win_prob = round(100 * finished_game.win_probability, 1)
-        team1_win_prob = round(100 - team0_win_prob, 1)
-        team0_str = f"{finished_game.team0_name} ({team0_win_prob}%): {team0_names}"
-        team1_str = f"{finished_game.team1_name} ({team1_win_prob}%): {team1_names}"
-
-        if finished_game.winning_team == 0:
-            output += f"\n**{team0_str}**"
-            output += f"\n{team1_str}"
-        elif finished_game.winning_team == 1:
-            output += f"\n{team0_str}"
-            output += f"\n**{team1_str}**"
-        else:
-            output += f"\n{team0_str}"
-            output += f"\n{team1_str}"
-        delta: timedelta = datetime.now(
-            timezone.utc
-        ) - finished_game.finished_at.replace(tzinfo=timezone.utc)
-        if delta.days > 0:
-            output += f"\n@ {delta.days} days ago\n"
-        elif delta.seconds > 3600:
-            hours_ago = delta.seconds // 3600
-            output += f"\n@ {hours_ago} hours ago\n"
-        else:
-            minutes_ago = delta.seconds // 60
-            output += f"\n@ {minutes_ago} minutes ago\n"
-        return output
-
-
 def in_progress_game_str(in_progress_game: InProgressGame, debug: bool = False) -> str:
     """
     Helper method to pretty print a finished game
@@ -940,45 +856,6 @@ def in_progress_game_str(in_progress_game: InProgressGame, debug: bool = False) 
             minutes_ago = delta.seconds // 60
             output += f"\n@ {minutes_ago} minutes ago\n"
         return output
-
-
-def is_in_game(player_id: int) -> bool:
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        return get_player_game(player_id, session) is not None
-
-
-def get_player_game(player_id: int, session=None) -> InProgressGame | None:
-    """
-    Find the game a player is currently in
-
-    :session: Pass in a session if you want to do something with the game that
-    gets returned
-    """
-    should_close = False
-    if not session:
-        should_close = True
-        session = (
-            Session()
-        )  # TODO: this session has the potential to not be closed, replace with context manager
-    ipg_player = (
-        session.query(InProgressGamePlayer)
-        .join(InProgressGame)
-        .filter(InProgressGamePlayer.player_id == player_id)
-        .first()
-    )
-    if ipg_player:
-        if should_close:
-            session.close()
-        return (
-            session.query(InProgressGame)
-            .filter(InProgressGame.id == ipg_player.in_progress_game_id)
-            .first()
-        )
-    else:
-        if should_close:
-            session.close()
-        return None
 
 
 # Commands start here
@@ -1642,178 +1519,9 @@ async def del_(ctx: Context, *args):
     session.close()
 
 
-@bot.command(usage="<player>")
-@commands.check(is_admin)
-async def delplayer(ctx: Context, member: Member, *args):
-    """
-    Admin command to delete player from all queues
-    """
-    message = ctx.message
-    session = ctx.session
-    queues: List[Queue] = (
-        session.query(Queue)
-        .join(QueuePlayer)
-        .filter(QueuePlayer.player_id == member.id)
-        .order_by(Queue.created_at.asc())
-        .all()
-    )  # type: ignore
-    for queue in queues:
-        session.query(QueuePlayer).filter(
-            QueuePlayer.queue_id == queue.id, QueuePlayer.player_id == member.id
-        ).delete()
-        # TODO: Test this part
-        queue_waitlist: QueueWaitlist | None = (
-            session.query(QueueWaitlist)
-            .filter(
-                QueueWaitlist.queue_id == queue.id,
-            )
-            .first()
-        )
-        if queue_waitlist:
-            session.query(QueueWaitlistPlayer).filter(
-                QueueWaitlistPlayer.player_id == member.id,
-                QueueWaitlistPlayer.queue_waitlist_id == queue_waitlist.id,
-            ).delete()
-
-    queue_statuses = []
-    queue: Queue
-    for queue in session.query(Queue).order_by(Queue.created_at.asc()).all():  # type: ignore
-        queue_players = (
-            session.query(QueuePlayer).filter(QueuePlayer.queue_id == queue.id).all()
-        )
-        queue_statuses.append(f"{queue.name} [{len(queue_players)}/{queue.size}]")
-
-    await send_message(
-        message.channel,
-        content=f"{escape_markdown(member.name)} removed from: {', '.join([queue.name for queue in queues])}",
-        embed_description=" ".join(queue_statuses),
-        colour=Colour.green(),
-    )
-    session.commit()
-
-
-@bot.command()
-@commands.check(is_admin)
-async def deletegame(ctx: Context, game_id: str):
-    message = ctx.message
-    session = ctx.session
-    finished_game: FinishedGame | None = (
-        session.query(FinishedGame)
-        .filter(FinishedGame.game_id.startswith(game_id))
-        .first()
-    )
-    if not finished_game:
-        await send_message(
-            message.channel,
-            embed_description=f"Could not find game: {game_id}",
-            colour=Colour.red(),
-        )
-        return
-    session.query(FinishedGamePlayer).filter(
-        FinishedGamePlayer.finished_game_id == finished_game.id
-    ).delete()
-    session.delete(finished_game)
-    session.commit()
-    await send_message(
-        message.channel,
-        embed_description=f"Game: **{finished_game.game_id}** deleted",
-        colour=Colour.green(),
-    )
-
-
 @bot.command()
 async def testleaderboard(ctx: Context):
     await print_leaderboard(ctx.channel)
-
-
-@bot.command()
-async def disableleaderboard(ctx: Context):
-    session = ctx.session
-    player = session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    player.leaderboard_enabled = False
-    session.commit()
-    await send_message(
-        ctx.message.channel,
-        embed_description="You are no longer visible on the leaderboard",
-        colour=Colour.blue(),
-    )
-
-
-@bot.command()
-async def disablestats(ctx: Context):
-    session = ctx.session
-    player = session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    player.stats_enabled = False
-    session.commit()
-    await send_message(
-        ctx.message.channel, embed_description="!stats disabled", colour=Colour.blue()
-    )
-
-
-@bot.command(usage="<game_id> <tie|be|ds>")
-@commands.check(is_admin)
-async def editgamewinner(ctx: Context, game_id: str, outcome: str):
-    message = ctx.message
-    session = ctx.session
-    game: FinishedGame | None = (
-        session.query(FinishedGame)
-        .filter(FinishedGame.game_id.startswith(game_id))
-        .first()
-    )
-    if not game:
-        await send_message(
-            message.channel,
-            embed_description=f"Could not find game: {game_id}",
-            colour=Colour.red(),
-        )
-        return
-    outcome = outcome.lower()
-    if outcome == "tie":
-        game.winning_team = -1
-    elif outcome == "be":
-        game.winning_team = 0
-    elif outcome == "ds":
-        game.winning_team = 1
-    else:
-        await send_message(
-            message.channel,
-            embed_description="Outcome must be tie, be, or ds",
-            colour=Colour.red(),
-        )
-        return
-
-    session.add(game)
-    session.commit()
-    await send_message(
-        message.channel,
-        embed_description=f"Game {game_id} outcome changed:\n\n"
-        + finished_game_str(game),
-        colour=Colour.green(),
-    )
-
-
-@bot.command()
-async def enableleaderboard(ctx: Context):
-    session = ctx.session
-    player = session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    player.leaderboard_enabled = True
-    session.commit()
-    await send_message(
-        ctx.message.channel,
-        embed_description="You are visible on the leaderboard",
-        colour=Colour.blue(),
-    )
-
-
-@bot.command()
-async def enablestats(ctx: Context):
-    session = ctx.session
-    player = session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    player.stats_enabled = True
-    session.commit()
-    await send_message(
-        ctx.message.channel, embed_description="!stats enabled", colour=Colour.blue()
-    )
 
 
 # @bot.command()
@@ -1936,36 +1644,6 @@ async def pug(ctx: Context):
 #         embed_description=f"Delay between games set to {RE_ADD_DELAY}",
 #         colour=Colour.green(),
 #     )
-
-
-@bot.command(usage="<true|false>")
-async def setmoveenabled(ctx: Context, enabled_option: bool = True):
-    session = ctx.session
-
-    if not config.ENABLE_VOICE_MOVE:
-        await send_message(
-            ctx.message.channel,
-            embed_description="Voice movement is disabled",
-            colour=Colour.red(),
-        )
-        return
-
-    player = session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    player.move_enabled = enabled_option
-    session.commit()
-
-    if enabled_option:
-        await send_message(
-            ctx.message.channel,
-            embed_description="Player moving enabled",
-            colour=Colour.blue(),
-        )
-    else:
-        await send_message(
-            ctx.message.channel,
-            embed_description="Player moving disabled",
-            colour=Colour.blue(),
-        )
 
 
 @bot.command()
