+++ discord_bots/main.py
@@ -16,6 +16,7 @@ from discord_bots.cogs.economy import EconomyCommands
 from discord_bots.cogs.in_progress_game import InProgressGameCommands
 from discord_bots.cogs.list import ListCommands
 from discord_bots.cogs.map import MapCommands
+from discord_bots.cogs.player import PlayerCommands
 from discord_bots.cogs.queue import QueueCommands
 from discord_bots.cogs.raffle import RaffleCommands
 from discord_bots.cogs.random import RandomCommands
@@ -80,17 +81,6 @@ async def on_app_command_error(
     interaction: Interaction, error: AppCommandError
 ) -> None:
     # TODO: provide more context about the error to the user
-    if interaction.response.is_done():
-        await interaction.followup.send(
-            embed=Embed(description="Oops! Something went wrong ☹️", color=Colour.red())
-        )
-    else:
-        # fallback case that responds to the interaction, since there always needs to be a response
-        await interaction.response.send_message(
-            embed=Embed(description="Oops! Something went wrong ☹️", color=Colour.red()),
-            ephemeral=True,
-        )
-
     if isinstance(error, errors.CheckFailure):
         return
     else:
@@ -100,6 +90,17 @@ async def on_app_command_error(
             )
         else:
             _log.error(f"[on_app_command_error]: {error}")
+    
+    if interaction.response.is_done():
+        await interaction.followup.send(
+            embed=Embed(description="Oops! Something went wrong ☹️", color=Colour.red())
+        )
+    else:
+        # fallback case that responds to the interaction, since there always needs to be a response
+        await interaction.response.send_message(
+            embed=Embed(description="Oops! Something went wrong ☹️", color=Colour.red()),
+            ephemeral=True,
+        )  
 
 
 @bot.event
@@ -261,6 +262,7 @@ async def setup():
     await bot.add_cog(InProgressGameCommands(bot))
     await bot.add_cog(ListCommands(bot))
     await bot.add_cog(MapCommands(bot))
+    await bot.add_cog(PlayerCommands(bot))
     await bot.add_cog(QueueCommands(bot))
     await bot.add_cog(RaffleCommands(bot))
     await bot.add_cog(RandomCommands(bot))
